interfaced all the stores and added factory to create device service specific ServiceObjects
