+++ src/test/java/org/edgexfoundry/device/store/ProfileStoreTest.java
@@ -33,6 +33,7 @@ import java.util.List;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.edgexfoundry.controller.DeviceProfileClient;
 import org.edgexfoundry.controller.ValueDescriptorClient;
+import org.edgexfoundry.device.store.impl.ProfileStoreImpl;
 import org.edgexfoundry.domain.common.ValueDescriptor;
 import org.edgexfoundry.domain.meta.Device;
 import org.edgexfoundry.domain.meta.DeviceProfile;
@@ -52,7 +53,7 @@ public class ProfileStoreTest {
 
 
   @InjectMocks
-  private ProfileStore profileStore;
+  private ProfileStoreImpl profileStore;
 
   @Mock
   private ValueDescriptorClient valueDescriptorClient;
